Successfully using textbox input to check weather

+++ Models/WeatherData.cs
@@ -18,15 +18,15 @@ namespace WeatherData
 	public class Main
 	{
 		public double temp { get; set; }
-		public int pressure { get; set; }
-		public int humidity { get; set; }
+		public double pressure { get; set; }
+		public double humidity { get; set; }
 		public double temp_min { get; set; }
 		public double temp_max { get; set; }
 	}
 	public class Wind
 	{
 		public double speed { get; set; }
-		public int deg { get; set; }
+		public double deg { get; set; }
 	}
 	public class Clouds
 	{
